Remove everything related to CTF sequences and variants
Those CTF types are not supported by barectf currently, yet some code
was written to support them at the configuration level. This is some
rather complex (dead) code to resolve field paths and validate those
types, but without anything to test it. In fact, the code exists to
parse a variant type, for example, but barectf fails with an error (not
supported) when you do so.

The code is still in this repository's history; let's add it back when
we can really test it.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>