barectf: add support for packet sequence number
Adds a new optional packet features object `seq-num-field-type`,
which generates the metadata field type name `packet_seq_num` when
defined (defaults to false).
Currently this is conditionally maintained in `struct barectf_ctx` as
a `uint32_t` type like `events_discarded`.
In the future this could be improved to use the config type defined
by the stream in `struct barectf_default_ctx` so that the types are aligned.

Signed-off-by: Jon Lamb <jon@auxon.io>