Fix: Pass C toolchain environments in tox.ini
CC is important if `cc' is not in PATH.

C_INCLUDE_PATH is important for the compiler to find the good headers,
e.g. <linux/limits.h>.

LIBRARY_PATH is important for the compiler to link against good
libraries, e.g. `crt1.o'.

This is mandatory for distributions that does not install
headers/libraries in standard locations that can be deduce by the
compiler, e.g. on Guix and NIX systems.

Note that this is intended for the C toolchain and not tests them self.
See https://gcc.gnu.org/onlinedocs/gcc/Environment-Variables.html.

Change-Id: I35dca0a929b6ba00ba9fe7381a5d5bddce901da1
Signed-off-by: Olivier Dion <odion@efficios.com>