tests: cleanup: Remove dead $include-replace tests
Context
-------

The $include-replace property appears to have been removed sometime in
the past. These 3 test cases are the only remaining references to this
property in the code base. The last reference to $include-replace in the
log is in commit 6839ffb (Use JSON schemas to validate the layout and
types of the YAML config) in `config_parse.py` at lines 2693-2694 where
the following comment referring to the property is removed:

    # In each object, only one of the $include and $include-replace
    # special properties is allowed.

All of this seems to point to this property no longer existing.

Changes
-------

Remove test cases referring to the $include-replace property.

Change-Id: I71d3293acf3518ebe5812db7fc9cc07ed933f64e
Signed-off-by: Erica Bugden <ebugden@efficios.com>