tests: fix: Run packet-set-buf tracing test
Context
-------

Currently, the packet-set-buf test is not run even though the test was
added as a regression test for a bug fix (0a5733a). The fix commit
message says:

    Adding a test with a custom platform calling
    barectf_packet_set_buf() with a `NULL` buffer after closing a
    completely full packet. Without this patch (in `barectf.c.j2`), this
    test leads to a segmentation fault.

It looks like this new test case was accidentally not added to the list
of valid test categories:

    * Sep 24, 2020 - Filter for valid test categories added (in 5b906ca)
    * Oct 6, 2020 - packet-set-buf test added, but not to valid
       categories (in 0a5733a)

Changes
-------

Add `packet-set-buf` to the valid test categories so the test will run.

Change-Id: I43a90bafe08897eaf107336fd8b629bad512b996
Signed-off-by: Erica Bugden <ebugden@efficios.com>